resolve namespace gamefranework problem , and no object generate problem (AKA <= can compare int and long)

+++ Source/Game/BtdClass/GameManager.cpp
@@ -116,7 +116,7 @@ namespace Btd
             }
         case Win:
             round++;
-            if (round >= Map.GetRounds().size())
+            if (round >= (int)Map.GetRounds().size())
             {
                 GameFlow = GameEnd;
             }
